reorganizing power model files
