+++ src/core/common.jl
@@ -156,22 +156,6 @@ end
 
 
 
-# NOTE, this function assumes all values are p.u. and angles are in radians
-function add_branch_parameters(data :: Dict{AbstractString,Any})
-    for branch in data["branch"]
-        r = branch["br_r"]
-        x = branch["br_x"]
-        tap_ratio = branch["tap"]
-        angle_shift = branch["shift"]
-
-        branch["g"] =  r/(x^2 + r^2)
-        branch["b"] = -x/(x^2 + r^2)
-        branch["tr"] = tap_ratio*cos(angle_shift)
-        branch["ti"] = tap_ratio*sin(angle_shift)
-    end
-end
-
-
 function standardize_cost_order(data :: Dict{AbstractString,Any})
     for gencost in data["gencost"]
         if gencost["model"] == 2 && length(gencost["cost"]) < 3
