+++ src/core/var.jl
