+++ src/prob/misc.jl
